refactor(data-loader): add kintone/upload.ts
