+++ packages/data-loader/src/controllers/export.ts
@@ -20,17 +20,18 @@ export const run = async (argv: RestAPIClientOptions & Options) => {
     ...restApiClientOptions
   } = argv;
   const apiClient = buildRestAPIClient(restApiClientOptions);
-  const records = await downloadRecords(apiClient, {
+  const records = await downloadRecords({
+    apiClient,
     app,
     condition,
     orderBy,
     attachmentsDir,
   });
   await printRecords({
+    apiClient,
     records,
     app,
     format,
-    apiClient,
     attachmentsDir,
   });
 };
